requirements.txt
aniso8601==8.0.0
appnope==0.1.0
asgiref==3.2.7
attrs==19.3.0
azure-common==1.1.25
azure-nspkg==3.0.2
azure-storage-blob==0.37.1
azure-storage-common==0.37.1
azure-storage-nspkg==3.1.0
backcall==0.1.0
boto==2.49.0
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
colour==0.1.5
cryptography==2.9.2
decorator==4.4.2
distro==1.5.0
Django==3.0.6
et-xmlfile==1.0.1
Flask==1.1.2
Flask-Cors==3.0.8
Flask-RESTful==0.3.8
futures==3.1.1
gevent==20.5.2
google-compute-engine==2.8.13
greenlet==0.4.15
gunicorn==20.0.4
idna==2.10
importlib-metadata==1.6.0
ipython==7.14.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jdcal==1.4.1
jedi==0.17.0
Jinja2==2.11.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mysqlclient==1.4.6
numpy==1.18.5
openpyxl==3.0.3
pandas==1.0.4
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.5
ptyprocess==0.6.0
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
SQLAlchemy==1.3.17
sqlparse==0.3.1
styleframe==3.0.1
traitlets==4.3.3
type-check==0.1.8
urllib3==1.25.9
wcwidth==0.1.9
Werkzeug==1.0.1
xlrd==1.2.0
XlsxWriter==1.2.9
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0